+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This change log adheres to standards from [Keep a CHANGELOG](http://keepachangelog.com).
 
 ## [Unreleased]
+- (none)
+
+## [1.2.0] - 2016-06-21
 ### Added
 - Added `exceptions` option to [`no-mutation`]
 - Added [`no-events`] rule
@@ -25,7 +28,8 @@ This change log adheres to standards from [Keep a CHANGELOG](http://keepachangel
 - Added [`no-this`] rule
 - Added [`no-throw`] rule
 
-[Unreleased]: https://github.com/jfmengels/eslint-plugin-fp/compare/v1.1.0...master
+[Unreleased]: https://github.com/jfmengels/eslint-plugin-fp/compare/v1.2.0...master
+[1.2.0]: https://github.com/jfmengels/eslint-plugin-fp/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/jfmengels/eslint-plugin-fp/compare/v1.0.0...v1.1.0
 
 [`no-class`]: ./docs/rules/no-class.md